% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cambell}
\alias{cambell}
\title{Leaf inclination distribution function
Ellipsoidal distribution function}
\usage{
cambell(ala, tx1, tx2)
}
\arguments{
\item{ala}{average leaf angle parameter}

\item{tx1}{angle in degree}

\item{tx2}{angle in degree}
}
\value{
angle fraction value
}
\description{
Leaf inclination distribution function
Ellipsoidal distribution function
}
