% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccrtm.R
\docType{package}
\name{ccrtm}
\alias{ccrtm}
\alias{ccrtm-package}
\title{ccrtm}
\description{
Coupled Chain Radiative Transfer Models.
A collection of radiative transfer models that can form a coupled chain to model
radiative transfer across multiple spatial scales from leaf to canopy to stand.
}
\author{
Marco D. Visser
}
