% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cdcum}
\alias{cdcum}
\title{Leaf inclination distribution function
cummulative lagden function from Wout Verhoef's dissertation
Extended here for any angle}
\usage{
cdcum(a, b, theta)
}
\arguments{
\item{a}{parameter}

\item{b}{parameter}

\item{theta}{angle in degrees}
}
\value{
angle fraction value
}
\description{
Leaf inclination distribution function
cummulative lagden function from Wout Verhoef's dissertation
Extended here for any angle
}
