% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cplateModel}
\alias{cplateModel}
\title{plate model}
\usage{
cplateModel(r12, t12, r21, t21, x, y, trans, N)
}
\arguments{
\item{(theta)}{Angle (in radians!)}

\item{(ref)}{refractive index}
}
\description{
plate model
}
\references{
Feret et al. (2008), PROSPECT-4 and 5: Advances in the Leaf Optical
Properties Model Separating Photosynthetic Pigments, Remote Sensing of
Environment

Stern F. (1964), Transmission of isotropic radiation across an
interface between two dielectrics, Appl. Opt., 3(1):111-113.

Allen W.A. (1973), Transmission of isotropic light across a
dielectric surface in two and three dimensions, J. Opt. Soc. Am.,
63(6):664-666.
}
