% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cReflTrans}
\alias{cReflTrans}
\title{Single layer reflection and transmission function 
4sail
\code{cReflTrans} calculates}
\usage{
cReflTrans(rho, tau, lai, att, m, sigb, ks, ko, sf, sb, vf, vb)
}
\arguments{
\item{rho}{numeric vector of particle reflectance}

\item{tau}{numeric vector of particle transmission}

\item{lai}{leaf area index}

\item{att}{}

\item{vb}{}
}
\value{
Returns a list of reflectances
}
\description{
Single layer reflection and transmission function 
4sail
\code{cReflTrans} calculates
}
\author{
Marco D. Visser
}
