% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cReflTransSingleLayer}
\alias{cReflTransSingleLayer}
\title{Single layer for two layer prospect
4sail2}
\usage{
cReflTransSingleLayer(rho, tau, lai, ks, ko, sdf, sdb, dof, dob, sob, sof,
  ddb, ddf)
}
\description{
Single layer for two layer prospect
4sail2
}
