% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccrtm.R
\docType{data}
\name{data_prospectd}
\alias{data_prospectd}
\title{refractive index and specific absorption coefficient for PROSPECT D}
\usage{
data(prospectd)
}
\description{
see http://teledetection.ipgp.jussieu.fr/prosail/ for more details
on the data.
}
\section{details}{

***********************************************************************
data_prospect5 (february, 25th 2008)
 The dataset contains the following labels (columns):
***********************************************************************
 \itemize{

\item [1] = wavelength (nm)
\item [2] = refractive index of leaf material (
 or the ratio of the velocity of light in a vacuum to its velocity in "leaf
medium").
\item [3] = specific absorption coefficient of chlorophyll (a+b) (cm2.microg-1)
\item [4] = specific absorption coefficient of carotenoids (cm2.microg-1)
\item [5] = specific absorption coefficient of brown pigments (arbitrary units)
\item [6] = specific absorption coefficient of water (cm-1)
\item [7] = specific absorption coefficient of dry matter (g.cm-1)
\item [8] = direct light
\item [9] = diffuse light
\item [10] = dry soil
\item [11] = wet soil

}
}

\section{references}{

Feret et al. (2008), PROSPECT-4 and 5: Advances in the Leaf Optical
Properties Model Separating Photosynthetic Pigments, Remote Sensing of
Environment
}

\keyword{datasets}
