% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{defaults.flim}
\alias{defaults.flim}
\title{d = stand density (d)
cd = crown diameter (cd)
h = mean crown height (h)
lai = leaf area index (lai)
alpha = light extinction coefficient (alpha)
tts = Solar zenith angle (tts)
tto = Observer zenith angle (tto)
psi = Sun-sensor azimuth angle (psi)}
\usage{
\method{defaults}{flim}(x, simple = TRUE)
}
\description{
d = stand density (d)
cd = crown diameter (cd)
h = mean crown height (h)
lai = leaf area index (lai)
alpha = light extinction coefficient (alpha)
tts = Solar zenith angle (tts)
tto = Observer zenith angle (tto)
psi = Sun-sensor azimuth angle (psi)
}
