% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{defaults.foursail}
\alias{defaults.foursail}
\title{Generate default values for the foursail model
      - psoil = linear mixing weight for 2 soil
                reflectance types
      - LAI   =  Leaf area index
      - TypeLidf= Type of leaf angle distribution. 
      - lidfa = Leaf angle distribution parameter a. 
      - lidfb = Leaf angle distribution parameter b.
      - hspot = Hotspot parameter
      - tts   = Solar zenith angle (degrees)
      - tto   = Observer zenith angle (degrees)
      - psi   = Relative azimuth angle (degrees)}
\usage{
\method{defaults}{foursail}(x)
}
\description{
Generate default values for the foursail model
      - psoil = linear mixing weight for 2 soil
                reflectance types
      - LAI   =  Leaf area index
      - TypeLidf= Type of leaf angle distribution. 
      - lidfa = Leaf angle distribution parameter a. 
      - lidfb = Leaf angle distribution parameter b.
      - hspot = Hotspot parameter
      - tts   = Solar zenith angle (degrees)
      - tto   = Observer zenith angle (degrees)
      - psi   = Relative azimuth angle (degrees)
}
