% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{defaults.foursail2b}
\alias{defaults.foursail2b}
\title{Generate default values for the foursail2b model 
      - psoil = psoil: Dry/Wet soil factor
      - LAI   =  Leaf area index
      - TypeLidf= Type of leaf angle distribution. 
      - lidfa = Leaf angle distribution parameter a. 
      - lidfb = Leaf angle distribution parameter b.
      - Cv    = vertical crown coverage fraction.
      - Zeta0 = tree shape factor (D/H).
      - fb    = fraction secondary particles (fraction "brown").
      - diss  = Canopy dissociation factor for secondary particles.
      - skyl  = linear mixing weight for diffuse & direct radiation.
                (ratio of diffuse to total incident radiation)
      - hspot = Hotspot parameter.
      - tts   = Solar zenith angle (degrees).
      - tto   = Observer zenith angle (degrees).
      - psi   = Relative azimuth angle (degrees).}
\usage{
\method{defaults}{foursail2b}(x)
}
\description{
Generate default values for the foursail2b model 
      - psoil = psoil: Dry/Wet soil factor
      - LAI   =  Leaf area index
      - TypeLidf= Type of leaf angle distribution. 
      - lidfa = Leaf angle distribution parameter a. 
      - lidfb = Leaf angle distribution parameter b.
      - Cv    = vertical crown coverage fraction.
      - Zeta0 = tree shape factor (D/H).
      - fb    = fraction secondary particles (fraction "brown").
      - diss  = Canopy dissociation factor for secondary particles.
      - skyl  = linear mixing weight for diffuse & direct radiation.
                (ratio of diffuse to total incident radiation)
      - hspot = Hotspot parameter.
      - tts   = Solar zenith angle (degrees).
      - tto   = Observer zenith angle (degrees).
      - psi   = Relative azimuth angle (degrees).
}
