% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{defaults.prospectd}
\alias{defaults.prospectd}
\title{get some default values for the prospect model 
      - N     = leaf structure parameter
      - Cab   = chlorophyll a+b content 
      - Car   = carotenoids content 
      - Canth = Leaf anthocyanin content
      - Cbrown= brown pigments content in arbitrary units
      - Cw    = equivalent water thickness
      - Cm    = dry matter content}
\usage{
\method{defaults}{prospectd}(x)
}
\description{
get some default values for the prospect model 
      - N     = leaf structure parameter
      - Cab   = chlorophyll a+b content 
      - Car   = carotenoids content 
      - Canth = Leaf anthocyanin content
      - Cbrown= brown pigments content in arbitrary units
      - Cw    = equivalent water thickness
      - Cm    = dry matter content
}
