% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect.R
\name{.prospect5}
\alias{.prospect5}
\title{Low level prospect D implementation}
\usage{
.prospect5(param, expected = c("n", "cab", "car", "cbrown", "cw", "cm"))
}
\description{
Low level prospect D implementation
}
