% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect.R
\name{.prospectd}
\alias{.prospectd}
\title{Low level prospect D implementation}
\usage{
.prospectd(param, expected = c("n", "cab", "car", "canth", "cbrown",
  "cw", "cm"))
}
\description{
Low level prospect D implementation
}
