% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flim.R
\name{flim}
\alias{flim}
\title{Forest Light Interaction Model (FLIM)}
\usage{
flim(Rc, Rg, To = NULL, Ts = NULL, params, area = 10000)
}
\arguments{
\item{Rc}{Canopy reflectance at infinite depth}

\item{Rg}{soil/background reflectance}

\item{To}{transmission in viewing direction}

\item{Ts}{transmission in sun direction}

\item{cd}{crown diameter (confounded with D)}

\item{D}{stand density (confounded with cd)}
}
\description{
Forests are assumed a discontinous mix of tree crowns
and gaps. Reflectance is modelled in terms of the
probabilty to observe either a gap (background)
or a tree crown. Both gaps and crowns may be shaded.
}
\details{
Confounded parameters pairs cannot be inversely
estimated, one of the pairs should be set to 1.
}
