% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertRTM.R
\name{invertRTM}
\alias{invertRTM}
\title{invert a requested RTM (internal function)}
\usage{
invertRTM(pars)
}
\arguments{
\item{pars}{the required parameters (vector or list),
and newdata}

\item{modReq}{model request object built in bRTM}
}
\value{
prediction from the requested model
}
\description{
List of aliases: prospect5, prospectd
}
