% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{KLd}
\alias{KLd}
\title{Kullback-Lieber divergence function
D(spec1 || spec2) = sum(spec1 * log(spec1 / spec2))}
\usage{
KLd(spec1, spec2)
}
\arguments{
\item{spec1}{spectral signal 1}

\item{spec2}{spectral signal 2 at identical wavelengths}
}
\value{
the KL divergence between the vector inputs
}
\description{
Kullback-Lieber divergence function
D(spec1 || spec2) = sum(spec1 * log(spec1 / spec2))
}
