% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.rtm.spectra}
\alias{plot.rtm.spectra}
\title{Plot RTM return spectra vs. wavelength}
\usage{
\method{plot}{rtm.spectra}(x, ...)
}
\arguments{
\item{x}{predictions from an RTM}

\item{\dots}{additional plot arguments}
}
\description{
Plot RTM return spectra vs. wavelength
}
