% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.rtm.inversion}
\alias{print.rtm.inversion}
\title{RTM inversion}
\usage{
\method{print}{rtm.inversion}(x, ...)
}
\arguments{
\item{x}{predictions from an RTM}

\item{\dots}{additional plot arguments}
}
\value{
prints the inverted parameters
}
\description{
RTM inversion
}
