% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.rtm.spectra}
\alias{print.rtm.spectra}
\title{Plot RTM return spectra vs. wavelength}
\usage{
\method{print}{rtm.spectra}(x, ...)
}
\arguments{
\item{x}{predictions from an RTM}

\item{\dots}{additional plot arguments}
}
\value{
prints the standard information from a simulated 
ccrtm spectra plot
}
\description{
Plot RTM return spectra vs. wavelength
}
