% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROSPECTR.R
\name{prospect_5BR}
\alias{prospect_5BR}
\title{The PROSPECT model of leaf optical properties (version 5B)}
\usage{
prospect_5BR(N = 1.5, Cab = 60, Car = 10, Cbrown = 0.01,
  Cw = 0.013, Cm = 0.003)
}
\arguments{
\item{(N)}{leaf structure parameter}

\item{(Cab)}{chlorophyll a+b content in ug/cm2}

\item{(Car)}{carotenoids content in ug/cm2}

\item{(Cbrown)}{brown pigments content in arbitrary units}

\item{(Cw)}{equivalent water thickness in g/cm2 or cm}

\item{(Cm)}{dry matter content in g/cm2}
}
\description{
Note from the original code; 
Jacquemoud S., Baret F. (1990), PROSPECT: a model of leaf optical
properties spectra, Remote Sens. Environ., 34:75-91.
Feret et al. (2008), PROSPECT-4 and 5: Advances in the Leaf Optical
Properties Model Separating Photosynthetic Pigments, Remote Sensing of
Environment, 112:3030-3043
The specific absorption coefficient corresponding to brown pigment is
provided by Frederic Baret (EMMAH, INRA Avignon, baret[at]avignon.inra.fr)
and used with his autorization.
***********************************************************************
}
