% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect.R
\name{prospectd}
\alias{prospectd}
\title{Prospect model version D}
\usage{
prospectd(param, data = NULL)
}
\arguments{
\item{param}{A named vector of PROSPECT parameters (note: program ignores case):
 \itemize{
\item [1] = leaf structure parameter (N)
\item [2] = chlorophyll a+b content in ug/cm2 (Cab)
\item [3] = carotenoids content in ug/cm2 (Car)
\item [4] = Leaf anthocyanin content (ug/cm2) (Canth)
\item [5] = brown pigments content in arbitrary units (Cbrown)
\item [6] = equivalent water thickness in g/cm2  (Cw)
\item [7] = leaf dry matter content in g/cm2 - lma - (Cm)
}}
}
\value{
spectra matrix with leaf reflectance and transmission 
for wavelengths 400 to 2500nm:
 \itemize{
\item [1] = leaf reflectance (rho)
\item [2] = leaf transmission (tau)
}
}
\description{
Prospect model version D
}
