% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foursail.R
\name{RTgeom}
\alias{RTgeom}
\title{RTgeom: correct rho and tau using SUITS factors}
\usage{
RTgeom(rho, tau, ddb, ddf, sdb, sdf, dob, dof, sob, sof)
}
\description{
RTgeom: correct rho and tau using SUITS factors
}
