% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRTM.R
\name{runRTM}
\alias{runRTM}
\title{run a requested RTM (internal function)}
\usage{
runRTM(modReq)
}
\arguments{
\item{modReq}{model request object built in fRTM}

\item{\dots}{additional plot arguments}
}
\value{
prediction from the requested model
}
\description{
run a requested RTM (internal function)
}
