% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foursail.R
\name{sail_BDRF}
\alias{sail_BDRF}
\title{The SAIL BDRF function}
\usage{
sail_BDRF(w, lai, sumint, tsstoo, rsoil, rdd, tdd, tsd, rsd, tdo, rdo, tss,
  too, rsod)
}
\arguments{
\item{w}{goemeric reflectance parameter}

\item{lai}{leaf area index}

\item{sumint}{exp int}

\item{tsstoo}{Bi-directional gap fraction}

\item{rsoil}{background reflectance}

\item{rdd}{Bi-hemispherical reflectance over all in & outgoing directions
(white-sky albedo).}

\item{tdd}{Bi-hemispherical transmittance (diffuse transmittance in all directions)}

\item{tsd}{Directional hemispherical transmittance for solar flux}

\item{rsd}{Directional hemispherical reflectance for solar flux (diffuse solar angle)}

\item{tdo}{Directional hemispherical transmittance (diffuse in viewing direction)}

\item{rdo}{Directional hemispherical reflectance in viewing direction}

\item{tss}{Direct transmittance of solar flux}

\item{too}{Direct transmittance in viewing direction}

\item{rsod}{Multi scattering contribution}
}
\description{
The SAIL BDRF function
}
