% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccrtm.R
\docType{data}
\name{soil}
\alias{soil}
\title{soil reflectance}
\usage{
data(soil)
}
\description{
soil reflectance
}
\section{details}{

***********************************************************************
 \itemize{
\item [1] = wet soil
\item [2] = dry soil 
}
}

\section{references}{

Feret et al. (2008), PROSPECT-4 and 5: Advances in the Leaf Optical
Properties Model Separating Photosynthetic Pigments, Remote Sensing of
Environment
}

\keyword{datasets}
