% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccrtm.R
\docType{data}
\name{solar}
\alias{solar}
\title{direct and diffuse light}
\usage{
data(solar)
}
\description{
direct and diffuse light
}
\section{details}{

***********************************************************************
 \itemize{
\item [1] = direct light
\item [2] = diffuse light
}
}

\section{references}{

Feret et al. (2008), PROSPECT-4 and 5: Advances in the Leaf Optical
Properties Model Separating Photosynthetic Pigments, Remote Sensing of
Environment
}

\keyword{datasets}
