% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foursail.R
\name{SUITS}
\alias{SUITS}
\title{SUITS function to calculate SUITS coefficients}
\usage{
SUITS(na, litab, lidv, tts, tto, cts, cto, psi, ctscto)
}
\description{
SUITS function to calculate SUITS coefficients
}
