% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.rtm.inversion}
\alias{summary.rtm.inversion}
\title{RTM inversion summary}
\usage{
\method{summary}{rtm.inversion}(x, ...)
}
\arguments{
\item{x}{predictions from an RTM}

\item{\dots}{additional plot arguments}
}
\value{
summarizes the inverted parameters
}
\description{
RTM inversion summary
}
