% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROSPECTR.R
\name{tav}
\alias{tav}
\title{tav function}
\usage{
tav(teta, ref)
}
\arguments{
\item{(teta)}{Angle in degrees}

\item{(ref)}{refractive index}
}
\description{
Stern F. (1964), Transmission of isotropic radiation across an
interface between two dielectrics, Appl. Opt., 3(1):111-113.
Allen W.A. (1973), Transmission of isotropic light across a
dielectric surface in two and three dimensions, J. Opt. Soc. Am.,
63(6):664-666.
}
\section{references}{

Feret et al. (2008), PROSPECT-4 and 5: Advances in the Leaf Optical
Properties Model Separating Photosynthetic Pigments, Remote Sensing of
Environment
}

