% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccrtm.R
\docType{package}
\name{ccrtm}
\alias{ccrtm}
\alias{ccrtm-package}
\title{ccrtm: Coupled Chain Radiative Transfer Models.}
\description{
A collection of radiative transfer models that can form
a coupled chain to model radiative transfer across multiple
spatial scales from leaf to canopy to stand.
}
\details{
Currently implemented models:

\itemize{

\item [1] = PROSPECT 5, 5B and D
\item [2] = FOURSAIL, and FOURSAIL2 
\item [3] = FLIM

}

Currently being tested or to be implemented models

\itemize{

\item [1] = LIBERTY, PROCOSINE
\item [2] = INFORM

}
}
\author{
Marco D. Visser
}
