% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{getDefaults}
\alias{getDefaults}
\title{S3- methods for Generate defaults settings and parameters
for all supported models.}
\usage{
getDefaults(model = NULL, ...)
}
\arguments{
\item{model}{a ccrtm formula or character vector of modelnames}

\item{\\dots}{not used.
(e.g. "prospect5")}
}
\value{
a data.frame with default model parameters
}
\description{
S3- methods for Generate defaults settings and parameters
for all supported models.
}
