% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRTM.R
\name{runRTM}
\alias{runRTM}
\title{run a requested RTM (internal function)}
\usage{
runRTM(pars)
}
\arguments{
\item{pars}{the required parameters (vector or list)}

\item{modReq}{model request object built in fRTM}
}
\value{
prediction from the requested model
}
\description{
List of aliases: prospect5, prospectd, prosail5,
prosaild, prosail2_55,prosail2_dd, prosail2_5d,
prosail2_d5, rtm.inform5, rtm.informd
}
