% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skyl.R
\name{skyl}
\alias{skyl}
\title{Sky light model}
\usage{
skyl(rddt, rsdt, rdot, rsot, Es, Ed, tts, skyl = NULL)
}
\arguments{
\item{rddt}{Bi-hemispherical reflectance}

\item{rsdt}{Directional-hemispherical reflectance for
solar incident flux}

\item{rdot}{Hemispherical-directional reflectance in viewing direction}

\item{rsot}{Bi-directional reflectance factor}

\item{Es}{Solar flux}

\item{Ed}{Diffuse flux}

\item{tts}{solar angle}

\item{skyl}{fraction diffuse}
}
\description{
simple atmospherical model
}
